requirements.txt
requests
azure-ai-projects
azure-identity
python-dotenv